added the combined redux method
