+++ src/components/book.js
@@ -1,19 +1,19 @@
-import React from 'react';
-import Proptypes from 'prop-types';
+import React from 'react'
+import Proptypes from 'prop-types'
 
 const Book = props => {
-  const { bookObject } = props;
+  const { bookObject } = props
   return (
     <tr>
       <td>{bookObject.id}</td>
       <td>{bookObject.title}</td>
       <td>{bookObject.category}</td>
     </tr>
-  );
-};
+  )
+}
 
 Book.propTypes = {
-  bookObject: Proptypes.object.isRequired,
-};
+  bookObject: Proptypes.object.isRequired
+}
 
-export default Book;
+export default Book
