+++ src/reducers/index.js
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux'
+import booksReducer from './bookReducer'
+
+const rootReducer = combineReducers({
+  book: booksReducer
+})
+
+export default rootReducer
