+++ src/store.js
@@ -1,6 +1,7 @@
-import { createStore } from 'redux';
-import booksReducer from './reducers/bookReducer';
+import { createStore } from 'redux'
+import booksReducer from './reducers/bookReducer'
+import rootReducer from './reducers/index'
 
-const store = createStore(booksReducer);
+const store = createStore(rootReducer)
 
-export default store;
+export default store
